telegram-summarizer/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
openai==1.58.1
aiohttp==3.10.11
youtube-transcript-api==0.6.3
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
pytz==2024.2
selenium==4.27.1
webdriver-manager==4.0.2
google-cloud-storage==2.18.2
